feat(pci-users): roles matrix mobile ref: DTCORE-1867
