+++ packages/manager/apps/pci-users/package.json
@@ -37,6 +37,7 @@
     "react-hook-form": "^7.50.1",
     "react-i18next": "^14.0.5",
     "react-router-dom": "^6.22.0",
+    "react-use": "^17.5.0",
     "zustand": "^4.5.0"
   },
   "devDependencies": {
