+++ packages/manager/apps/pci-users/src/layout.ts
@@ -1,0 +1,7 @@
+export enum BreakPoints {
+  XS = 0,
+  SM = 540,
+  MD = 720,
+  LG = 960,
+  XL = 1140,
+}
